Refactor and simply code, undo change which create problem
